Don't pass 'embed-bitcode'/'lto' while building proc-macros
cargo is usually good about not passing `RUSTFLAGS` to build-time
dependencies which need to be built for the host platform rather
than the target platform, but for some reason it does not do so
for `proc-macro` crates.

Thus, we have to manually drop the `embed-bitcode` and `lto`
arguments when calling rustc in our existing rustc wrapper.