Hard-code dependent crate feature flags
When we generate bindings, we generate bindings across the API
surface of whatever features we had set. Thus, any generated
bindings are specific to the features we had set and it doesn't
make sense to have features on the `lightning-c-bindings` crate as
setting the wrong one will simply lead to a compilation failure.

Instead, here we hard-code the features in the dependencies section
of `Cargo.toml`.

Sadly, because we want to take an optional `core2` dependency, we
can't quite fully drop the `lightning-c-bindings` features
entirely, but after upgrading to rust-bitcoin 0.32 we should be
able to.