Ensure we comply with the non-NULL req for `slice::from_raw_parts`
When building a slice with `slice::from_raw_parts`, we're required
to pass in a non-NULL pointer even if the length is 0. While we did
this in most cases, we weren't always careful about it. Here we fix
the issue by using a simple wrapper which passes a dummy pointer
instead.